backend/requirements.txt
google-cloud-aiplatform
google-genai
livekit-agents
livekit-plugins-google
pandas
openpyxl
python-dotenv
fastapi
uvicorn[standard]
python-multipart
